back/Hairo_Back/requirements.txt
aioredis==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==3.6.4.0
boto3==1.17.44
botocore==1.20.112
cachetools==5.3.3
celery==5.1.2
channels==3.0.3
channels-redis==3.2.0
charset-normalizer==3.3.2
click==7.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contourpy==1.1.1
cycler==0.12.1
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.41.0
django-core==1.4.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-debug-toolbar==4.2.0
django-redis==4.12.1
django-storages==1.11.1
django-webpack-loader==3.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
fonttools==4.43.1
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
hiredis==2.3.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jmespath==0.10.0
kiwisolver==1.4.5
kombu==5.3.7
matplotlib==3.8.0
msgpack==1.0.8
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
Pillow==8.3.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.21.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.1
python3-openid==3.2.0
redis==3.5.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.3.7
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.3
vue-cal== 4.8.1
